requirements.txt
requests
beautifulsoup4
fake-useragent
dnspython
aiogram
python-dotenv
python-whois